+++ src/playground/Playground.tsx
@@ -6,11 +6,11 @@ import DropdownMenu from '@/components/common/Dropdown/DropdownMenu';
 import Button from '@/components/common/Button/Button';
 import Divider from '@/components/common/Divider/Divider';
 import Modal from '@/components/common/Modal/Modal';
-import PageHeader from '@/components/common/PageHeader/PageHeader';
-import PageHeaderLegacy from '@/components/common/PageHeader/PageHeaderLegacy';
+import PageHeader from '@/components/common/Pageheader/PageHeader';
+import PageHeaderLegacy from '@/components/common/Pageheader/PageHeaderLegacy';
 import Pagination from '@/components/common/Pagination/Pagination';
-import TextField from '@/components/common/text/TextField';
-import Textarea from '@/components/common/text/Textarea';
+import TextField from '@/components/common/Text/TextField';
+import Textarea from '@/components/common/Text/Textarea';
 import TextBox from '@/components/common/Textbox/TextBox';
 import Toast from '@/components/common/Toast/Toast';
 import Sidebar from '@/layout/Sidebar';
